Split stage tests into package
