+++ tests/unit/test_stages/utils.py
@@ -1,0 +1,7 @@
+import mock
+
+USER_ID = 1
+OTHER_USER_ID = 2
+GROUP_ID = 10
+OTHER_GROUP_ID = 11
+patch_obj = mock.patch.object  # pylint:disable=invalid-name
